refactor(theme): refactor markers and tests
