+++ packages/react-ui/lib/theming/ThemeFactory.ts
@@ -2,22 +2,12 @@ import { DefaultThemeInternal } from '../../internal/themes/DefaultTheme';
 import { isNonNullable } from '../utils';
 
 import { Theme, ThemeIn } from './Theme';
-import { findPropertyDescriptor, Markers, REACT_UI_THEME_MARKERS } from './ThemeHelpers';
+import { findPropertyDescriptor, REACT_UI_THEME_MARKERS } from './ThemeHelpers';
 
 export class ThemeFactory {
-  public static create<T extends unknown>(theme: ThemeIn & T): Readonly<Theme & T>;
-  public static create<T extends unknown>(theme: ThemeIn & T, baseTheme: Theme): Readonly<Theme & T>;
-  public static create<T extends unknown>(theme: ThemeIn & T, baseTheme: Theme, markers: Markers): Readonly<Theme & T>;
-  public static create<T extends unknown>(theme: ThemeIn & T, markers: Markers): Readonly<Theme & T>;
-  public static create<T extends unknown>(
-    theme: ThemeIn & T,
-    baseThemeOrMarkers?: Theme | Markers,
-    markers?: Markers,
-  ): Readonly<Theme & T> {
-    const base = baseThemeOrMarkers && !Array.isArray(baseThemeOrMarkers) ? baseThemeOrMarkers : DefaultThemeInternal;
-    // eslint-disable-next-line no-nested-ternary
-    const _markers = Array.isArray(baseThemeOrMarkers) ? baseThemeOrMarkers : Array.isArray(markers) ? markers : [];
-    return this.constructTheme(base, theme, _markers);
+  public static create<T extends unknown>(theme: ThemeIn & T, baseTheme?: Theme): Readonly<Theme & T> {
+    const base = baseTheme || DefaultThemeInternal;
+    return this.constructTheme(base, theme);
   }
 
   public static overrideDefaultTheme(theme: Theme) {
@@ -50,7 +40,7 @@ export class ThemeFactory {
     return keys.sort();
   }
 
-  private static constructTheme(base: Theme, theme: ThemeIn, markers: Markers) {
+  private static constructTheme(base: Theme, theme: ThemeIn) {
     const newTheme = Object.create(base);
     Object.keys(theme).forEach((propName) => {
       const descriptor = Object.getOwnPropertyDescriptor(theme, propName);
@@ -58,8 +48,6 @@ export class ThemeFactory {
         Object.defineProperty(newTheme, propName, descriptor);
       }
     });
-
-    markers.forEach((marker) => marker(newTheme));
 
     return Object.freeze(newTheme);
   }
